Lemma and Indexes realisation ready

+++ src/main/java/searchengine/services/impl/IndexingServiceImpl.java
@@ -2,9 +2,11 @@ package searchengine.services.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.lucene.morphology.russian.RussianLuceneMorphology;
+import org.jsoup.HttpStatusException;
 import org.jsoup.Jsoup;
 import org.springframework.stereotype.Service;
+import searchengine.dto.indexing.results.IndexingResult;
+import searchengine.dto.indexing.results.IndexingResultSuccess;
 import searchengine.model.Index;
 import searchengine.model.Lemma;
 import searchengine.repository.IndexRepository;
@@ -12,7 +14,7 @@ import searchengine.repository.LemmaRepository;
 import searchengine.utils.PageProcessor;
 import searchengine.config.Site;
 import searchengine.config.SitesList;
-import searchengine.dto.indexing.IndexingResult;
+import searchengine.dto.indexing.results.IndexingResultFail;
 import searchengine.model.Page;
 import searchengine.model.Status;
 import searchengine.repository.PageRepository;
@@ -21,10 +23,10 @@ import searchengine.services.IndexingService;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.concurrent.ForkJoinPool;
 
 @Service
@@ -40,40 +42,55 @@ public class IndexingServiceImpl implements IndexingService {
     private final SitesList sitesList;
 
     private static boolean indexingStopped = false;
+    private static boolean indexingNow = false;
 
     @Override
     public IndexingResult getIndexingResult() {
-
         indexingStopped = false;
-
+        System.out.println(indexingNow);
+        if (indexingNow) {
+            return new IndexingResultFail("Индексация уже запущена");
+        }
+        indexingNow = true;
         log.info("Method started");
 
 
         List<Site> siteList = sites.getSites();
         for (Site siteConfig : siteList) {
+
             searchengine.model.Site siteDB = configSiteToModelSite(siteConfig);
             for (searchengine.model.Site siteModel : siteRepository.findAll()) {
                 if (siteModel.getUrl().equals(siteConfig.getUrl())) {
                     siteRepository.delete(siteModel);
                 }
             }
-            siteDB.setStatus(Status.INDEXING);
             siteDB.setStatusTime(LocalDateTime.now());
+
+            try {
+                Jsoup.connect(siteDB.getUrl()).execute();
+            } catch (HttpStatusException httpStatusException) {
+                siteDB.setStatus(Status.FAILED);
+                setErrors(siteDB, httpStatusException.getStatusCode());
+                siteRepository.save(siteDB);
+                continue;
+            } catch (IOException exception) {
+                siteDB.setStatus(Status.FAILED);
+                exception.printStackTrace();
+                siteRepository.save(siteDB);
+                continue;
+            }
+
+            siteDB.setStatus(Status.INDEXING);
             siteRepository.save(siteDB);
 
 
             log.info("Site saved to DB");
 
         }
-
+        System.out.println(indexingNow);
         setPagesInDB();
 
-        IndexingResult indexingResult = new IndexingResult();
-        indexingResult.setResult(true);
-
-//        PageProcessor.getLinkSet().clear();
-
-        return indexingResult;
+        return new IndexingResultSuccess();
     }
 
     @Override
@@ -88,8 +105,8 @@ public class IndexingServiceImpl implements IndexingService {
             }
         }
         if (!siteIsFromConfig) {
-            //Данная страница за пределами сайтов
-            return null;
+            return new IndexingResultFail("Данная страница находится за пределами " +
+                    "сайтов, указанных в конфигурационном файле");
         }
 
         searchengine.model.Site siteDB = configSiteToModelSite(configSite);
@@ -102,11 +119,27 @@ public class IndexingServiceImpl implements IndexingService {
                 break;
             }
         }
+
+        siteDB.setStatusTime(LocalDateTime.now());
+
+        try {
+            Jsoup.connect(siteDB.getUrl()).execute();
+        } catch (HttpStatusException httpStatusException) {
+            siteDB.setStatus(Status.FAILED);
+            setErrors(siteDB, httpStatusException.getStatusCode());
+            siteRepository.save(siteDB);
+            return new IndexingResultFail("Невозможно получить доступ к сайту. Номер ошибки: " + httpStatusException.getStatusCode());
+        } catch (IOException exception) {
+            siteDB.setStatus(Status.FAILED);
+            exception.printStackTrace();
+            siteRepository.save(siteDB);
+            return new IndexingResultFail("Невозможно получить доступ к сайту.");
+        }
+
         if (!siteAlreadyInRepository) {
             siteDB.setStatus(Status.INDEXING);
         }
 
-        siteDB.setStatusTime(LocalDateTime.now());
         siteRepository.save(siteDB);
 
         boolean isPageAlreadyInDB = false;
@@ -120,37 +153,73 @@ public class IndexingServiceImpl implements IndexingService {
         page.setPath(path);
         page.setSite(siteDB);
 
+        Page pageFromDB = new Page();
+
         for (Page pageDB : pageRepository.findAll()) {
             if (page.equals(pageDB)) {
-                pageRepository.delete(pageDB);
+                pageFromDB = pageDB;
                 isPageAlreadyInDB = true;
                 break;
             }
         }
 
+        if (!addPageLemmasAndIndexes(page, pageUrl, isPageAlreadyInDB, pageFromDB, siteDB)) {
+            return new IndexingResultFail("Невозможно получить доступ к странице");
+        }
+
+        siteDB.setStatusTime(LocalDateTime.now());
+        siteRepository.save(siteDB);
+
+        return new IndexingResultSuccess();
+    }
+
+    @Override
+    public IndexingResult stopIndexing() {
+
+        if (!indexingNow) {
+            return new IndexingResultFail("Индексация не запущена");
+        }
+
+        indexingNow = false;
+        indexingStopped = true;
+
+        return new IndexingResultSuccess();
+    }
+
+    private boolean addPageLemmasAndIndexes(Page page, String pageUrl, boolean isPageAlreadyInDB, Page pageFromDB, searchengine.model.Site siteDB) {
         try {
             page.setCode(Jsoup.connect(pageUrl).execute().statusCode());
+
             String getHtml = Jsoup.connect(pageUrl).get().html();
             page.setContent(getHtml);
-
-            pageRepository.save(page);
-
-            //Lemmas code
-            LemmaServiceImpl lemmaServiceImpl = new LemmaServiceImpl(lemmaRepository, pageRepository);
+            LemmaServiceImpl lemmaServiceImpl = new LemmaServiceImpl();
             HashMap<String, Integer> lemmas = lemmaServiceImpl.collectLemmas(getHtml);
+
+            if (isPageAlreadyInDB) {
+
+                List<Index> indexesFromThePageIndexRepository = indexRepository.findAllByPage(pageFromDB);
+                List<Lemma> lemmasFromThePageIndexRepository = new ArrayList<>();
+                for (Index index : indexesFromThePageIndexRepository) {
+                    lemmasFromThePageIndexRepository.add(index.getLemma());
+                }
+
+                for (Lemma lemmaOnThePage : lemmasFromThePageIndexRepository) {
+
+                    lemmaOnThePage.setFrequency(lemmaOnThePage.getFrequency() - 1);
+                    if (lemmaOnThePage.getFrequency() == 0) {
+                        lemmaRepository.delete(lemmaOnThePage);
+                    }
+                }
+
+                pageRepository.delete(pageFromDB);
+            }
+            pageRepository.save(page);
 
             for (Map.Entry<String, Integer> entry : lemmas.entrySet()) {
                 boolean isLemmaAlreadyInDB = false;
                 Lemma newLemma = new Lemma();
                 for (Lemma lemma : lemmaRepository.findAll()) {
                     if (lemma.getLemma().equals(entry.getKey())) {
-
-
-                        //TODO: Нужно удалить старые леммы
-                        if (isPageAlreadyInDB) {
-                            lemma.setFrequency(lemma.getFrequency() - 1);
-                        }
-
                         lemma.setFrequency(lemma.getFrequency() + 1);
                         lemmaRepository.save(lemma);
                         isLemmaAlreadyInDB = true;
@@ -170,34 +239,21 @@ public class IndexingServiceImpl implements IndexingService {
                 index.setRank(entry.getValue());
                 indexRepository.save(index);
             }
-
+        } catch (HttpStatusException httpStatusException) {
+            page.setCode(httpStatusException.getStatusCode());
+            pageRepository.save(page);
+            return false;
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            e.printStackTrace();
+            return false;
         }
-
-//        for (Page pageDB : pageRepository.findAll()) {
-//            if (pageDB.equals(page)) {
-//                return;
-//            }
-//        }
-
-        siteDB.setStatusTime(LocalDateTime.now());
-        siteRepository.save(siteDB);
-
-        return null;
-    }
-
-    @Override
-    public IndexingResult stopIndexing() {
-        indexingStopped = true;
-
-        IndexingResult indexingResult = new IndexingResult();
-        indexingResult.setResult(true);
-        return indexingResult;
+        return true;
     }
 
 
     private void setPagesInDB() {
+
+        int sitesIndexed = 0;
 
         for (Site siteConfig : sites.getSites()) {
             new Thread(() -> {
@@ -223,11 +279,20 @@ public class IndexingServiceImpl implements IndexingService {
                     siteRepository.save(siteDB);
 
                 } else {
-                    siteDB.setStatus(Status.INDEXED);
-
-                    siteRepository.save(siteDB);
+                    if (!siteDB.getStatus().equals(Status.FAILED)) {
+                        siteDB.setStatus(Status.INDEXED);
+                        siteRepository.save(siteDB);
+                    }
                 }
             }).start();
+        }
+        for (searchengine.model.Site siteDB : siteRepository.findAll()) {
+            if (siteDB.getStatus().equals(Status.INDEXED)) {
+                sitesIndexed++;
+            }
+        }
+        if (sitesIndexed == siteRepository.count()) {
+            indexingNow = false;
         }
     }
 
@@ -238,7 +303,24 @@ public class IndexingServiceImpl implements IndexingService {
         return siteDB;
     }
 
+    private void setErrors(searchengine.model.Site site, int errorCode) {
+        switch (errorCode) {
+            case 400 -> site.setLastError("Код ошибки: " + errorCode + " - Bad Request");
+            case 401 -> site.setLastError("Код ошибки: " + errorCode + " - Unauthorized Error");
+            case 403 -> site.setLastError("Код ошибки: " + errorCode + " - Forbidden");
+            case 404 -> site.setLastError("Код ошибки: " + errorCode + " - Not Found");
+            case 405 -> site.setLastError("Код ошибки: " + errorCode + " - Method Not Allowed");
+            case 500 -> site.setLastError("Код ошибки: " + errorCode + " - Internal Server Error");
+            default -> site.setLastError("Невозможно получить доступ к сайту, код ошибки " + errorCode);
+        }
+    }
+
+
     public static boolean isIndexingStopped() {
         return indexingStopped;
+    }
+
+    public static boolean isIndexingNow() {
+        return indexingNow;
     }
 }
